fix(watched-users): compact list items + half-star render + race condition fix
- item_movie_watched_user: avatar 36→30dp, username 15→13sp, icons 14→12dp, padding 8→6dp
- MovieWatchedUsersAdapter: star size 17→12dp, gap 0→1dp
- MovieApiController: rating cast (int)→(float) fix half-star data loss
- MovieActionsHelper: add userInteractedWithLike/Watchlist flags to prevent initial DB load from overwriting user actions

+++ UKOM/Moview_backend/app/Http/Controllers/Api/MovieApiController.php
@@ -621,7 +621,7 @@ class MovieApiController extends Controller
                         'username' => $item->username,
                         'profile_photo' => $profilePhoto,
                     ],
-                    'rating' => $item->rating !== null ? (int) $item->rating : null,
+                    'rating' => $item->rating !== null ? (float) $item->rating : null,
                     'review_id' => $item->review_id !== null ? (int) $item->review_id : null,
                     'has_like' => (bool) $item->has_like,
                     'has_review' => (bool) $item->has_review,
